SantityCheck

+++ Assets/Scripts/UI/CurrencyUI.cs
@@ -7,7 +7,7 @@ public class CurrencyUI : MonoBehaviour
 {
     public TextMeshProUGUI currencyText;
 
-    public void updateCurrencyDisplayed(int currency)
+    public void updateCurrencyDisplayed(long currency)
     {
         currencyText.text = currency.ToString();
     }
